reorg wind converter models fixtures and mark test types
