+++ h2integrate/converters/wind/test/test_atb_wind_with_pysam.py
@@ -1,73 +1,13 @@
 import pytest
 import openmdao.api as om
-from pytest import fixture
 
 from h2integrate.converters.wind.wind_pysam import PYSAMWindPlantPerformanceModel
 from h2integrate.converters.wind.atb_wind_cost import ATBWindPlantCostModel
 from h2integrate.resource.wind.nrel_developer_wtk_api import WTKNRELDeveloperAPIWindResource
 
 
-@fixture
-def wind_resource_config():
-    wind_resource_dict = {
-        "latitude": 35.2018863,
-        "longitude": -101.945027,
-        "resource_year": 2012,
-    }
-    return wind_resource_dict
-
-
-@fixture
-def plant_config():
-    site_config = {
-        "latitude": 35.2018863,
-        "longitude": -101.945027,
-    }
-    plant_dict = {
-        "plant_life": 30,
-        "simulation": {"n_timesteps": 8760, "dt": 3600, "start_time": "01/01 00:30:00"},
-    }
-
-    d = {"site": site_config, "plant": plant_dict}
-    return d
-
-
-@fixture
-def wind_plant_config():
-    layout_config = {
-        "layout_mode": "basicgrid",
-        "layout_options": {
-            "row_D_spacing": 5.0,
-            "turbine_D_spacing": 5.0,
-            "rotation_angle_deg": 0.0,
-            "row_phase_offset": 0.0,
-            "layout_shape": "square",
-        },
-    }
-    pysam_config = {
-        "Farm": {
-            "wind_farm_wake_model": 0,
-        },
-        "Losses": {
-            "ops_strategies_loss": 10.0,
-        },
-    }
-    design_config = {
-        "num_turbines": 50,
-        "hub_height": 115,
-        "rotor_diameter": 170,
-        "turbine_rating_kw": 6000,
-        "create_model_from": "default",
-        "config_name": "WindPowerSingleOwner",
-        "pysam_options": pysam_config,
-        "layout": layout_config,
-    }
-    return design_config
-
-
-def test_wind_plant_costs_with_pysam(
-    wind_resource_config, plant_config, wind_plant_config, subtests
-):
+@pytest.mark.regression
+def test_wind_plant_costs_with_pysam(plant_config_wtk, wind_plant_config, subtests):
     cost_dict = {
         "capex_per_kW": 1000,  # overnight capital cost
         "opex_per_kW_per_year": 5,  # fixed operations and maintenance expenses
@@ -82,22 +22,22 @@ def test_wind_plant_costs_with_pysam(
 
     prob = om.Problem()
 
-    plant_config["site"].update({"resources": {"wind_resource": wind_resource_config}})
-
     wind_resource = WTKNRELDeveloperAPIWindResource(
-        plant_config=plant_config,
-        resource_config=wind_resource_config,
+        plant_config=plant_config_wtk,
+        resource_config=plant_config_wtk["site"]["resource"]["wind_resource"][
+            "resource_parameters"
+        ],
         driver_config={},
     )
 
     wind_plant = PYSAMWindPlantPerformanceModel(
-        plant_config=plant_config,
+        plant_config=plant_config_wtk,
         tech_config=tech_config_dict,
         driver_config={},
     )
 
     wind_cost = ATBWindPlantCostModel(
-        plant_config=plant_config,
+        plant_config=plant_config_wtk,
         tech_config=tech_config_dict,
         driver_config={},
     )
